+++ server/src/app/views/product.html
